Add Check to delete.lua
delete.lua checks now, if it has the permission to delete a file.

+++ src/main/resources/assets/computercraft/lua/rom/programs/delete.lua
@@ -9,7 +9,11 @@ for i = 1, args.n do
     local files = fs.find(shell.resolve(args[i]))
     if #files > 0 then
         for n, file in ipairs(files) do
-            fs.delete(file)
+            if fs.isReadOnly( file ) or file == "" then
+                printError( "/" .. file .. ": Access denied")
+            else
+                fs.delete(file)
+            end
         end
     else
         printError(args[i] .. ": No matching files")
